+++ include/alpaka/acc/AccCpuOmp2Blocks.hpp
@@ -33,7 +33,7 @@
 // Implementation details.
 #include "alpaka/acc/Tag.hpp"
 #include "alpaka/core/ClipCast.hpp"
-#include "alpaka/core/Concepts.hpp"
+#include "alpaka/core/Interface.hpp"
 #include "alpaka/dev/DevCpu.hpp"
 
 #include <limits>
