+++ include/alpaka/acc/AccCpuSerial.hpp
@@ -31,7 +31,7 @@
 
 // Implementation details.
 #include "alpaka/acc/Tag.hpp"
-#include "alpaka/core/Concepts.hpp"
+#include "alpaka/core/Interface.hpp"
 #include "alpaka/dev/DevCpu.hpp"
 
 #include <memory>
