+++ include/alpaka/acc/AccCpuThreads.hpp
@@ -32,7 +32,7 @@
 #include "alpaka/acc/Tag.hpp"
 #include "alpaka/core/BoostPredef.hpp"
 #include "alpaka/core/ClipCast.hpp"
-#include "alpaka/core/Concepts.hpp"
+#include "alpaka/core/Interface.hpp"
 #include "alpaka/dev/DevCpu.hpp"
 
 #include <memory>
