+++ include/alpaka/acc/AccGenericSycl.hpp
@@ -33,7 +33,7 @@
 // Implementation details.
 #include "alpaka/core/BoostPredef.hpp"
 #include "alpaka/core/ClipCast.hpp"
-#include "alpaka/core/Concepts.hpp"
+#include "alpaka/core/Interface.hpp"
 #include "alpaka/core/Sycl.hpp"
 
 #include <cstddef>
