+++ include/alpaka/acc/AccGpuUniformCudaHipRt.hpp
@@ -30,8 +30,8 @@
 
 // Implementation details.
 #include "alpaka/core/ClipCast.hpp"
-#include "alpaka/core/Concepts.hpp"
 #include "alpaka/core/Cuda.hpp"
+#include "alpaka/core/Interface.hpp"
 #include "alpaka/dev/DevUniformCudaHipRt.hpp"
 
 #include <typeinfo>
