+++ include/alpaka/atomic/Traits.hpp
@@ -6,7 +6,7 @@
 
 #include "alpaka/atomic/Op.hpp"
 #include "alpaka/core/Common.hpp"
-#include "alpaka/core/Concepts.hpp"
+#include "alpaka/core/Interface.hpp"
 #include "alpaka/core/Positioning.hpp"
 
 #include <type_traits>
