+++ include/alpaka/block/shared/dyn/BlockSharedMemDynUniformCudaHipBuiltIn.hpp
@@ -6,7 +6,7 @@
 
 #include "alpaka/block/shared/dyn/Traits.hpp"
 #include "alpaka/core/BoostPredef.hpp"
-#include "alpaka/core/Concepts.hpp"
+#include "alpaka/core/Interface.hpp"
 
 #include <cstddef>
 #include <type_traits>
