+++ include/alpaka/block/sync/BlockSyncUniformCudaHipBuiltIn.hpp
@@ -6,7 +6,7 @@
 
 #include "alpaka/block/sync/Traits.hpp"
 #include "alpaka/core/BoostPredef.hpp"
-#include "alpaka/core/Concepts.hpp"
+#include "alpaka/core/Interface.hpp"
 
 #if defined(ALPAKA_ACC_GPU_CUDA_ENABLED) || defined(ALPAKA_ACC_GPU_HIP_ENABLED)
