+++ include/alpaka/dev/DevUniformCudaHipRt.hpp
@@ -6,9 +6,9 @@
 #pragma once
 
 #include "alpaka/core/ApiCudaRt.hpp"
-#include "alpaka/core/Concepts.hpp"
 #include "alpaka/core/Cuda.hpp"
 #include "alpaka/core/Hip.hpp"
+#include "alpaka/core/Interface.hpp"
 #include "alpaka/dev/Traits.hpp"
 #include "alpaka/dev/common/QueueRegistry.hpp"
 #include "alpaka/mem/buf/Traits.hpp"
