+++ include/alpaka/idx/gb/IdxGbGenericSycl.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "alpaka/core/Concepts.hpp"
+#include "alpaka/core/Interface.hpp"
 #include "alpaka/core/Positioning.hpp"
 #include "alpaka/core/Sycl.hpp"
 #include "alpaka/idx/Traits.hpp"
