+++ include/alpaka/math/MathGenericSycl.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "alpaka/core/Concepts.hpp"
+#include "alpaka/core/Interface.hpp"
 #include "alpaka/math/Complex.hpp"
 #include "alpaka/math/Traits.hpp"
