+++ include/alpaka/math/MathUniformCudaHipBuiltIn.hpp
@@ -6,9 +6,9 @@
 #pragma once
 
 #include "alpaka/core/BoostPredef.hpp"
-#include "alpaka/core/Concepts.hpp"
 #include "alpaka/core/CudaHipCommon.hpp"
 #include "alpaka/core/Decay.hpp"
+#include "alpaka/core/Interface.hpp"
 #include "alpaka/core/UniformCudaHip.hpp"
 #include "alpaka/core/Unreachable.hpp"
 #include "alpaka/math/Complex.hpp"
