+++ include/alpaka/mem/alloc/AllocCpuAligned.hpp
@@ -6,7 +6,7 @@
 
 #include "alpaka/core/AlignedAlloc.hpp"
 #include "alpaka/core/Common.hpp"
-#include "alpaka/core/Concepts.hpp"
+#include "alpaka/core/Interface.hpp"
 #include "alpaka/dev/cpu/SysInfo.hpp"
 #include "alpaka/mem/alloc/Traits.hpp"
