+++ include/alpaka/platform/PlatformUniformCudaHipRt.hpp
@@ -5,9 +5,9 @@
 
 #pragma once
 
-#include "alpaka/core/Concepts.hpp"
 #include "alpaka/core/Cuda.hpp"
 #include "alpaka/core/Hip.hpp"
+#include "alpaka/core/Interface.hpp"
 #include "alpaka/dev/DevUniformCudaHipRt.hpp"
 #include "alpaka/dev/Traits.hpp"
