+++ include/alpaka/warp/WarpUniformCudaHipBuiltIn.hpp
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "alpaka/core/BoostPredef.hpp"
-#include "alpaka/core/Concepts.hpp"
+#include "alpaka/core/Interface.hpp"
 #include "alpaka/warp/Traits.hpp"
 
 #include <cstdint>
